PR-005: move internal implementation into src/internal tree
