+++ src/config/settings.rs
@@ -1,7 +1,6 @@
 mod app_settings;
-mod embedding_repository_settings;
-mod vector_memory_repository_settings;
 
+pub(crate) use crate::internal::config::settings::{
+    EmbeddingRepositorySettings, VectorMemoryRepositorySettings,
+};
 pub use app_settings::Settings;
-pub(crate) use embedding_repository_settings::EmbeddingRepositorySettings;
-pub(crate) use vector_memory_repository_settings::VectorMemoryRepositorySettings;
