+++ src/infrastructures.rs
@@ -1,8 +1,3 @@
-// Module declarations
-pub(crate) mod external_services {
-    mod openai_embedding_repository;
-    mod qdrant_vector_memory_repository;
-
-    pub(crate) use openai_embedding_repository::OpenAIEmbeddingRepository;
-    pub(crate) use qdrant_vector_memory_repository::QdrantVectorMemoryRepository;
-}
+// Legacy module kept temporarily for transition.
+// Implementation types have moved under `crate::internal::infrastructures`.
+pub(crate) use crate::internal::infrastructures::*;
