+++ src/internal/config/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod settings;
