+++ src/internal/config/settings/embedding_repository_settings.rs
@@ -1,4 +1,4 @@
-use crate::models::vector::EmbeddingModel;
+use crate::internal::models::vector::EmbeddingModel;
 
 /// Settings for configuring the OpenAI Embedding Repository.
 ///
