+++ src/internal/config/settings/mod.rs
@@ -1,0 +1,5 @@
+mod embedding_repository_settings;
+mod vector_memory_repository_settings;
+
+pub(crate) use embedding_repository_settings::EmbeddingRepositorySettings;
+pub(crate) use vector_memory_repository_settings::VectorMemoryRepositorySettings;
