+++ src/internal/config/settings/vector_memory_repository_settings.rs
@@ -1,4 +1,4 @@
-use crate::models::vector::EmbeddingModel;
+use crate::internal::models::vector::EmbeddingModel;
 
 /// Configuration for VectorMemoryRepository.
 ///
