+++ src/internal/mod.rs
@@ -1,1 +1,4 @@
-// Internal implementation modules will be moved under this tree in PR-005.
+pub(crate) mod config;
+pub(crate) mod infrastructures;
+pub(crate) mod models;
+pub(crate) mod repositories;
