+++ src/internal/models/memory/mod.rs
@@ -1,0 +1,5 @@
+mod memory_entry;
+mod scored_memory_entry;
+
+pub(crate) use memory_entry::MemoryEntry;
+pub(crate) use scored_memory_entry::ScoredMemoryEntry;
