+++ src/internal/models/memory/scored_memory_entry.rs
@@ -1,5 +1,5 @@
 use crate::api::types::ScoredMemory;
-use crate::models::memory::MemoryEntry;
+use crate::internal::models::memory::MemoryEntry;
 
 #[derive(Debug, Clone)]
 pub(crate) struct ScoredMemoryEntry {
