+++ src/internal/models/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod memory;
+pub(crate) mod vector;
