+++ src/internal/models/vector/embedding_model.rs
