+++ src/internal/models/vector/mod.rs
@@ -1,0 +1,5 @@
+mod embedding_model;
+mod vector_metadata;
+
+pub(crate) use embedding_model::EmbeddingModel;
+pub(crate) use vector_metadata::VectorMetadata;
