+++ src/internal/models/vector/vector_metadata.rs
