+++ src/internal/repositories/mod.rs
@@ -1,0 +1,5 @@
+mod memory_repository;
+mod vector_memory_repository;
+
+pub(crate) use memory_repository::MemoryRepository;
+pub(crate) use vector_memory_repository::VectorMemoryRepository;
