+++ src/internal/repositories/vector_memory_repository.rs
@@ -3,7 +3,7 @@ use uuid::Uuid;
 
 use crate::api::types::MemoryFilters;
 use crate::errors::CustomError;
-use crate::models::memory::{MemoryEntry, ScoredMemoryEntry};
+use crate::internal::models::memory::{MemoryEntry, ScoredMemoryEntry};
 
 /// Repository trait for storing and retrieving memories using a vector database.
 ///
